@@ -2,15 +2,16 @@
 #include "lexer/Source.h"
 #include "lexer/Scanner.h"
 #include "parser/Parser.h"
+#include <memory>
 
 int main(int argc, char** argv) {
-    Source* source;
-    Scanner* scanner;
-    Parser* parser;
+    std::unique_ptr<Source> source;
+    std::unique_ptr<Scanner> scanner;
+    std::unique_ptr<Parser> parser;
 
     if (argc > 1)
         try {
-            source = new Source(argv[1]);
+            source = std::make_unique<Source>(argv[1]);
         } catch (const std::runtime_error& error) {
             std::cout << error.what() << std::endl;
             return -1;
@@ -18,8 +19,8 @@ int main(int argc, char** argv) {
     else
         return -1;
 
-    scanner = new Scanner(source);
-    parser = new Parser(scanner);
+    scanner = std::make_unique<Scanner>(std::move(source));
+    parser = std::make_unique<Parser>(std::move(scanner));
 
     try {
         parser->parseProgram();
